Relationships and Configuration

+++ server/mainframe/src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { Task } from './task.entity';
+import { User } from 'src/auth/user.entity';
 import { TaskStatus } from './task-status.enum'
 import { v1 as uuid} from 'uuid'
 import { CreateTaskDto } from './dto/create-task.dto';
@@ -14,34 +15,30 @@ export class TasksService {
         private taskRepository: TaskRepository
     ) {}
 
-    async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTask(filterDto)
+    async getTasks(filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
+        return this.taskRepository.getTask(filterDto, user)
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id)
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id } })
         if (!found) {
             throw new NotFoundException(`Task with id ${id} not found.`)
         }
         return found
     }
 
-    async getTasksWithFilters(filterDto: GetTaskFilterDto): Promise<Task[]> {
-        return await this.getTasks(filterDto)
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user)
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto)
-    }
-
-    async deleteTaskById(id: number): Promise<Task> {
-        const task = await this.getTaskById(id)
-        this.taskRepository.delete(id)
+    async deleteTaskById(id: number, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user)
+        this.taskRepository.delete({ id, userId: user.id })
         return task
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task>{
-        const task = await this.getTaskById(id)
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task>{
+        const task = await this.getTaskById(id, user)
         task.status = status
         await task.save()
         return task
